# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ePKL Backend API",
        "description": "Backend for ePKL mobile app",
        "contact": {
            "name": "Muhamamd Haikal",
            "email": "realityinaship@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "description": "Register a new user.",
                "operationId": "Register",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_lengkap": {
                                        "description": "Full name of the user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email address of the user",
                                        "type": "string"
                                    },
                                    "nis": {
                                        "description": "User's NIS",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Log in the user.",
                "operationId": "Login",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email address of the user",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Log out the user.",
                "operationId": "Logout",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/send-email-confirmation/{email}": {
            "post": {
                "tags": [
                    "Email Confirmation"
                ],
                "summary": "Send Email Confirmation Request",
                "description": "Send an email confirmation request to the specified email address.",
                "operationId": "SendEmailConfirmationRequest",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email address of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Email confirmation request sent"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/activate-email/{token}": {
            "get": {
                "tags": [
                    "Email Confirmation"
                ],
                "summary": "Activate Email",
                "description": "Activate the email using the provided token.",
                "operationId": "ActivateEmail",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token received in the email confirmation request",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email successfully activated"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/reset-password-form/{token}": {
            "get": {
                "tags": [
                    "Forgot Password"
                ],
                "summary": "Reset Password Form",
                "description": "Retrieve the reset password form using the provided token.",
                "operationId": "ResetPasswordForm",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token received in the password reset request",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reset password form retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/send-forgot-password-request/{email}": {
            "post": {
                "tags": [
                    "Forgot Password"
                ],
                "summary": "Send Forgot Password Request",
                "description": "Send a forgot password request to the specified email address.",
                "operationId": "SendForgotPasswordRequest",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email address of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Forgot password request sent"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/reset-password": {
            "post": {
                "tags": [
                    "Forgot Password"
                ],
                "summary": "Reset Password",
                "description": "Reset the password using the provided token and new password.",
                "operationId": "ResetPassword",
                "requestBody": {
                    "description": "Reset password details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/get-token-data/{token}": {
            "get": {
                "tags": [
                    "Forgot Password"
                ],
                "summary": "Get Token Data",
                "description": "Retrieve the data associated with the provided token.",
                "operationId": "GetTokenData",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token received in the password reset request",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reset token data retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}